clean up load_file(), restore ICE r load address functionality, use extern
Some clean-ups for the 1.38 release

1. The z80core doesn't know anything about page sizes, move the
   page start and size shifting from load_file() etc. into memory.c
   of altairsim, cromemcosim, and imsaisim where it belongs.
2. Remove parsing stuff from load_file().
3. Add a do_load() function and restore "r file,addr" functionality
   for Mostek binary files in the ICE.
4. Use "extern" for external function declaration.
5. Some files where missing a end-of-file '\n'.

+++ z80core/sim0.c
@@ -75,7 +75,7 @@
 static void init_cpu(void);
 static void save_core(void);
 int load_core(void);
-extern int load_file(char *, BYTE, WORD);
+extern int load_file(char *, WORD, int);
 extern void int_on(void), int_off(void), mon(void);
 extern void init_io(void), exit_io(void);
 extern int exatoi(char *);
